PAE-1669: rename stream port bulkAppendEvents to appendEvents, drop singular appendEvent
The all-or-nothing batch append is the only append primitive now that the
summary-log and PRN write paths route through it, so rename bulkAppendEvents to
appendEvents and remove the singular appendEvent method, its adapter
implementations, and its contract block.

Delete the orphaned append-to-stream.js (and its tests) — the last non-contract
caller of the singular method — after confirming it has no live importer. Fold
the singular contract's unique behaviours (full-field persistence, error
payloads, cross-partition slots) into the renamed appendEvents contract, and
migrate every remaining singular caller to the array form.
