+++ src/waste-balances/application/current-waste-balance.test.js
@@ -3,21 +3,29 @@ import { describe, it, expect } from 'vitest'
 import { createInMemoryStreamRepository } from '../repository/stream-inmemory.js'
 import { STREAM_EVENT_KIND } from '../repository/stream-schema.js'
 import { buildStreamEvent } from '../repository/stream-test-data.js'
-import { appendToStream } from './append-to-stream.js'
 import { currentWasteBalance } from './current-waste-balance.js'
 
-const createdBy = { id: 'user-1', name: 'Test User' }
 const partition = {
   organisationId: 'org-1',
   registrationId: 'reg-1',
   accreditationId: 'acc-1'
 }
-const context = (repository) => ({
-  repository,
-  registrationId: 'reg-1',
-  accreditationId: 'acc-1',
-  organisationId: 'org-1'
-})
+
+const submissionEvent = (number, creditTotal) =>
+  buildStreamEvent({
+    number,
+    kind: STREAM_EVENT_KIND.SUMMARY_LOG_SUBMITTED,
+    payload: { summaryLogId: `log-${number}`, creditTotal },
+    closingBalance: { amount: creditTotal, availableAmount: creditTotal }
+  })
+
+const prnCreatedEvent = (number, amount, closingBalance) =>
+  buildStreamEvent({
+    number,
+    kind: STREAM_EVENT_KIND.PRN_CREATED,
+    payload: { prnId: 'prn-1', amount },
+    closingBalance
+  })
 
 describe('currentWasteBalance', () => {
   it('returns null for an empty partition', async () => {
@@ -29,22 +37,10 @@ describe('currentWasteBalance', () => {
   it('resolves balance, head, and credit total from the stream', async () => {
     const repository = createInMemoryStreamRepository()()
 
-    await appendToStream(
-      { ...context(repository), expectedHead: 0 },
-      {
-        kind: STREAM_EVENT_KIND.SUMMARY_LOG_SUBMITTED,
-        payload: { summaryLogId: 'log-1', creditTotal: 1000 },
-        createdBy
-      }
-    )
-    await appendToStream(
-      { ...context(repository), expectedHead: 1 },
-      {
-        kind: STREAM_EVENT_KIND.PRN_CREATED,
-        payload: { prnId: 'prn-1', amount: 300 },
-        createdBy
-      }
-    )
+    await repository.appendEvents([
+      submissionEvent(1, 1000),
+      prnCreatedEvent(2, 300, { amount: 1000, availableAmount: 700 })
+    ])
 
     const balance = await currentWasteBalance(repository, partition)
 
@@ -62,22 +58,10 @@ describe('currentWasteBalance', () => {
   it('carries the latest credit total when several submissions precede a PRN', async () => {
     const repository = createInMemoryStreamRepository()()
 
-    await appendToStream(
-      { ...context(repository), expectedHead: 0 },
-      {
-        kind: STREAM_EVENT_KIND.SUMMARY_LOG_SUBMITTED,
-        payload: { summaryLogId: 'log-1', creditTotal: 1000 },
-        createdBy
-      }
-    )
-    await appendToStream(
-      { ...context(repository), expectedHead: 1 },
-      {
-        kind: STREAM_EVENT_KIND.SUMMARY_LOG_SUBMITTED,
-        payload: { summaryLogId: 'log-2', creditTotal: 2500 },
-        createdBy
-      }
-    )
+    await repository.appendEvents([
+      submissionEvent(1, 1000),
+      submissionEvent(2, 2500)
+    ])
 
     const balance = await currentWasteBalance(repository, partition)
 
@@ -88,22 +72,10 @@ describe('currentWasteBalance', () => {
   it('keeps the credit-total base at the latest submission when a PRN follows', async () => {
     const repository = createInMemoryStreamRepository()()
 
-    await appendToStream(
-      { ...context(repository), expectedHead: 0 },
-      {
-        kind: STREAM_EVENT_KIND.SUMMARY_LOG_SUBMITTED,
-        payload: { summaryLogId: 'log-1', creditTotal: 1000 },
-        createdBy
-      }
-    )
-    await appendToStream(
-      { ...context(repository), expectedHead: 1 },
-      {
-        kind: STREAM_EVENT_KIND.PRN_CREATED,
-        payload: { prnId: 'prn-1', amount: 300 },
-        createdBy
-      }
-    )
+    await repository.appendEvents([
+      submissionEvent(1, 1000),
+      prnCreatedEvent(2, 300, { amount: 1000, availableAmount: 700 })
+    ])
 
     const balance = await currentWasteBalance(repository, partition)
 
@@ -112,7 +84,7 @@ describe('currentWasteBalance', () => {
 
   it('reports a zero credit total for a partition with no submission event', async () => {
     const repository = createInMemoryStreamRepository()()
-    await repository.appendEvent(
+    await repository.appendEvents([
       buildStreamEvent({
         registrationId: 'reg-1',
         accreditationId: 'acc-1',
@@ -121,7 +93,7 @@ describe('currentWasteBalance', () => {
         payload: { prnId: 'prn-1', amount: 0 },
         closingBalance: { amount: 0, availableAmount: 0 }
       })
-    )
+    ])
 
     const balance = await currentWasteBalance(repository, partition)
