+++ src/waste-balances/repository/stream-contract/appendEvents.contract.js
@@ -3,8 +3,8 @@ import { describe, beforeEach, expect } from 'vitest'
 import { buildStreamEvent } from '../stream-test-data.js'
 import { StreamSequenceError, StreamSlotConflictError } from '../stream-port.js'
 
-export const testBulkAppendEventsBehaviour = (it) => {
-  describe('bulkAppendEvents', () => {
+export const testAppendEventsBehaviour = (it) => {
+  describe('appendEvents', () => {
     let repository
 
     beforeEach(
@@ -16,6 +16,26 @@ export const testBulkAppendEventsBehaviour = (it) => {
         repository = await streamRepository()
       }
     )
+
+    it('persists an event and returns the stored event with an id', async () => {
+      const event = buildStreamEvent({
+        registrationId: 'reg-append',
+        accreditationId: 'acc-append',
+        number: 1
+      })
+
+      const [stored] = await repository.appendEvents([event])
+
+      expect(stored.id).toEqual(expect.any(String))
+      expect(stored.id).not.toBe('')
+      expect(stored.registrationId).toBe('reg-append')
+      expect(stored.accreditationId).toBe('acc-append')
+      expect(stored.number).toBe(1)
+      expect(stored.kind).toBe(event.kind)
+      expect(stored.payload).toEqual(event.payload)
+      expect(stored.openingBalance).toEqual(event.openingBalance)
+      expect(stored.closingBalance).toEqual(event.closingBalance)
+    })
 
     it('inserts multiple events and returns stored events with ids', async () => {
       const events = [
@@ -32,7 +52,7 @@ export const testBulkAppendEventsBehaviour = (it) => {
         })
       ]
 
-      const stored = await repository.bulkAppendEvents(events)
+      const stored = await repository.appendEvents(events)
 
       expect(stored).toHaveLength(2)
       expect(stored[0].id).toEqual(expect.any(String))
@@ -42,13 +62,13 @@ export const testBulkAppendEventsBehaviour = (it) => {
     })
 
     it('throws StreamSequenceError when first event does not start at currentMax + 1', async () => {
-      await repository.appendEvent(
+      await repository.appendEvents([
         buildStreamEvent({
           registrationId: 'reg-seq',
           accreditationId: 'acc-seq',
           number: 1
         })
-      )
+      ])
 
       const events = [
         buildStreamEvent({
@@ -59,7 +79,7 @@ export const testBulkAppendEventsBehaviour = (it) => {
         })
       ]
 
-      await expect(repository.bulkAppendEvents(events)).rejects.toBeInstanceOf(
+      await expect(repository.appendEvents(events)).rejects.toBeInstanceOf(
         StreamSequenceError
       )
     })
@@ -73,19 +93,45 @@ export const testBulkAppendEventsBehaviour = (it) => {
         })
       ]
 
-      await expect(repository.bulkAppendEvents(events)).rejects.toBeInstanceOf(
+      await expect(repository.appendEvents(events)).rejects.toBeInstanceOf(
         StreamSequenceError
       )
     })
 
+    it('StreamSequenceError carries the provided and expected numbers', async () => {
+      await repository.appendEvents([
+        buildStreamEvent({
+          registrationId: 'reg-seq-err',
+          accreditationId: 'acc-seq-err',
+          number: 1
+        })
+      ])
+
+      await expect(
+        repository.appendEvents([
+          buildStreamEvent({
+            registrationId: 'reg-seq-err',
+            accreditationId: 'acc-seq-err',
+            number: 5,
+            payload: { summaryLogId: 'log-5', creditTotal: 500 }
+          })
+        ])
+      ).rejects.toMatchObject({
+        registrationId: 'reg-seq-err',
+        accreditationId: 'acc-seq-err',
+        providedNumber: 5,
+        expectedNumber: 2
+      })
+    })
+
     it('throws StreamSlotConflictError when the starting slot is already occupied', async () => {
-      await repository.appendEvent(
+      await repository.appendEvents([
         buildStreamEvent({
           registrationId: 'reg-occupied',
           accreditationId: 'acc-occupied',
           number: 1
         })
-      )
+      ])
 
       const events = [
         buildStreamEvent({
@@ -96,9 +142,34 @@ export const testBulkAppendEventsBehaviour = (it) => {
         })
       ]
 
-      await expect(repository.bulkAppendEvents(events)).rejects.toBeInstanceOf(
+      await expect(repository.appendEvents(events)).rejects.toBeInstanceOf(
         StreamSlotConflictError
       )
+    })
+
+    it('StreamSlotConflictError carries the partition identity and slot number', async () => {
+      await repository.appendEvents([
+        buildStreamEvent({
+          registrationId: 'reg-slot-err',
+          accreditationId: 'acc-slot-err',
+          number: 1
+        })
+      ])
+
+      await expect(
+        repository.appendEvents([
+          buildStreamEvent({
+            registrationId: 'reg-slot-err',
+            accreditationId: 'acc-slot-err',
+            number: 1,
+            payload: { summaryLogId: 'log-different', creditTotal: 200 }
+          })
+        ])
+      ).rejects.toMatchObject({
+        registrationId: 'reg-slot-err',
+        accreditationId: 'acc-slot-err',
+        slotNumber: 1
+      })
     })
 
     it('throws StreamSequenceError when events are not sequentially numbered', async () => {
@@ -116,13 +187,33 @@ export const testBulkAppendEventsBehaviour = (it) => {
         })
       ]
 
-      await expect(repository.bulkAppendEvents(events)).rejects.toBeInstanceOf(
+      await expect(repository.appendEvents(events)).rejects.toBeInstanceOf(
         StreamSequenceError
       )
     })
 
+    it('allows the same slot number across different partitions', async () => {
+      await repository.appendEvents([
+        buildStreamEvent({
+          registrationId: 'reg-a',
+          accreditationId: 'acc-a',
+          number: 1
+        })
+      ])
+
+      await expect(
+        repository.appendEvents([
+          buildStreamEvent({
+            registrationId: 'reg-b',
+            accreditationId: 'acc-b',
+            number: 1
+          })
+        ])
+      ).resolves.toBeDefined()
+    })
+
     it('is a no-op for an empty array', async () => {
-      const stored = await repository.bulkAppendEvents([])
+      const stored = await repository.appendEvents([])
 
       expect(stored).toEqual([])
     })
@@ -142,7 +233,7 @@ export const testBulkAppendEventsBehaviour = (it) => {
         })
       ]
 
-      await repository.bulkAppendEvents(events)
+      await repository.appendEvents(events)
 
       const latest = await repository.findLatestByPartition(
         'reg-vis',
