+++ src/waste-balances/repository/stream-mongodb.test.js
@@ -104,7 +104,7 @@ describe('MongoDB stream repository', () => {
   }) => {
     const database = mongoClient.db(DATABASE_NAME)
     const repository = (await createMongoStreamRepository(database))()
-    expect(repository.appendEvent).toBeTypeOf('function')
+    expect(repository.appendEvents).toBeTypeOf('function')
     expect(repository.findLatestByPartition).toBeTypeOf('function')
     expect(repository.findLatestByPartitionAndKind).toBeTypeOf('function')
     expect(repository.findEventsByPrnIdAfter).toBeTypeOf('function')
@@ -115,70 +115,7 @@ describe('MongoDB stream repository', () => {
     testStreamRepositoryContract(it)
   })
 
-  describe('appendEvent error translation', () => {
-    it('rethrows non-conflict MongoDB errors unchanged', async () => {
-      const upstream = new Error('connection lost')
-      const stubCollection = {
-        createIndex: () => Promise.resolve(),
-        findOne: () => Promise.resolve(null),
-        insertOne: () => Promise.reject(upstream)
-      }
-      const stubDb = { collection: () => stubCollection }
-
-      const repository = (
-        await createMongoStreamRepository(/** @type {*} */ (stubDb))
-      )()
-
-      await expect(
-        repository.appendEvent(buildStreamEvent({ number: 1 }))
-      ).rejects.toBe(upstream)
-    })
-
-    it('rethrows E11000 with unrecognised keyPattern as the raw error', async () => {
-      const mongoError = Object.assign(new Error('E11000'), {
-        code: 11000,
-        keyPattern: { unknownField: 1 }
-      })
-      const stubCollection = {
-        createIndex: () => Promise.resolve(),
-        findOne: () => Promise.resolve(null),
-        insertOne: () => Promise.reject(mongoError)
-      }
-      const stubDb = { collection: () => stubCollection }
-
-      const repository = (
-        await createMongoStreamRepository(/** @type {*} */ (stubDb))
-      )()
-
-      await expect(
-        repository.appendEvent(buildStreamEvent({ number: 1 }))
-      ).rejects.toBe(mongoError)
-    })
-
-    it('classifies E11000 from writeErrors array', async () => {
-      const mongoError = Object.assign(new Error('E11000'), {
-        writeErrors: [{ code: 11000, keyPattern: { number: 1 } }]
-      })
-      const stubCollection = {
-        createIndex: () => Promise.resolve(),
-        findOne: () => Promise.resolve(null),
-        insertOne: () => Promise.reject(mongoError)
-      }
-      const stubDb = { collection: () => stubCollection }
-
-      const repository = (
-        await createMongoStreamRepository(/** @type {*} */ (stubDb))
-      )()
-
-      await expect(
-        repository.appendEvent(buildStreamEvent({ number: 1 }))
-      ).rejects.toMatchObject({
-        name: 'StreamSlotConflictError'
-      })
-    })
-  })
-
-  describe('bulkAppendEvents error translation', () => {
+  describe('appendEvents error translation', () => {
     it('rethrows non-conflict MongoDB errors unchanged', async () => {
       const upstream = new Error('connection lost')
       const stubCollection = {
@@ -193,8 +130,29 @@ describe('MongoDB stream repository', () => {
       )()
 
       await expect(
-        repository.bulkAppendEvents([buildStreamEvent({ number: 1 })])
+        repository.appendEvents([buildStreamEvent({ number: 1 })])
       ).rejects.toBe(upstream)
+    })
+
+    it('rethrows E11000 with unrecognised keyPattern as the raw error', async () => {
+      const mongoError = Object.assign(new Error('E11000'), {
+        code: 11000,
+        keyPattern: { unknownField: 1 }
+      })
+      const stubCollection = {
+        createIndex: () => Promise.resolve(),
+        findOne: () => Promise.resolve(null),
+        insertMany: () => Promise.reject(mongoError)
+      }
+      const stubDb = { collection: () => stubCollection }
+
+      const repository = (
+        await createMongoStreamRepository(/** @type {*} */ (stubDb))
+      )()
+
+      await expect(
+        repository.appendEvents([buildStreamEvent({ number: 1 })])
+      ).rejects.toBe(mongoError)
     })
 
     it('classifies a slot conflict raised when inserting the batch', async () => {
@@ -213,7 +171,7 @@ describe('MongoDB stream repository', () => {
       )()
 
       await expect(
-        repository.bulkAppendEvents([buildStreamEvent({ number: 1 })])
+        repository.appendEvents([buildStreamEvent({ number: 1 })])
       ).rejects.toMatchObject({
         name: 'StreamSlotConflictError'
       })
