+++ src/waste-balances/repository/stream-port.contract.js
@@ -1,17 +1,15 @@
-import { testAppendEventBehaviour } from './stream-contract/appendEvent.contract.js'
+import { testAppendEventsBehaviour } from './stream-contract/appendEvents.contract.js'
 import { testFindLatestByPartitionBehaviour } from './stream-contract/findLatestByPartition.contract.js'
 import { testFindLatestByPartitionAndKindBehaviour } from './stream-contract/findLatestByPartitionAndKind.contract.js'
 import { testFindEventsByPrnIdAfterBehaviour } from './stream-contract/findEventsByPrnIdAfter.contract.js'
 import { testFindAllByPartitionBehaviour } from './stream-contract/findAllByPartition.contract.js'
 import { testDeleteByPartitionBehaviour } from './stream-contract/deleteByPartition.contract.js'
-import { testBulkAppendEventsBehaviour } from './stream-contract/bulkAppendEvents.contract.js'
 
 export const testStreamRepositoryContract = (it) => {
-  testAppendEventBehaviour(it)
+  testAppendEventsBehaviour(it)
   testFindLatestByPartitionBehaviour(it)
   testFindLatestByPartitionAndKindBehaviour(it)
   testFindEventsByPrnIdAfterBehaviour(it)
   testFindAllByPartitionBehaviour(it)
   testDeleteByPartitionBehaviour(it)
-  testBulkAppendEventsBehaviour(it)
 }
